requirements.txt
streamlit
python-dotenv
langchain_community
langchain_text_splitters
langchain_chroma
langchain_google_genai
langchain_classic
langchain_core
google-genai
pypdf
fastembed